+++ src/DemoFile.Test/Integration/DemoParserIntegrationTest.cs
@@ -1,16 +1,34 @@
-﻿using System.Text;
-using System.Text.Json;
-
-namespace DemoFile.Test.Integration;
+﻿namespace DemoFile.Test.Integration;
 
 [TestFixture]
 public class DemoParserIntegrationTest
 {
     [Test]
-    public async Task Parse()
+    public async Task ReadAll()
     {
         var demo = new DemoParser();
-        await demo.Start(GotvCompetitiveProtocol13963, default);
+        await demo.ReadAllAsync(GotvCompetitiveProtocol13963, default);
+        Assert.That(demo.CurrentDemoTick.Value, Is.EqualTo(217866));
+    }
+
+    [Test]
+    public async Task ByTick()
+    {
+        // Arrange
+        var demo = new DemoParser();
+        var tick = demo.CurrentDemoTick;
+
+        // Act
+        await demo.StartReadingAsync(GotvCompetitiveProtocol13963, default);
+        while (await demo.MoveNextAsync(default))
+        {
+            // Tick is monotonic
+            Assert.That(demo.CurrentDemoTick.Value, Is.GreaterThanOrEqualTo(tick.Value));
+            tick = demo.CurrentDemoTick;
+        }
+
+        // Assert
+        Assert.That(demo.CurrentDemoTick.Value, Is.EqualTo(217866));
     }
 
     private static readonly KeyValuePair<string, Stream>[] CompatibilityCases =
@@ -20,16 +38,16 @@ public class DemoParserIntegrationTest
     };
 
     [TestCaseSource(nameof(CompatibilityCases))]
-    public async Task Parse_Compatibility(KeyValuePair<string, Stream> testCase)
+    public async Task ReadAll_Compatibility(KeyValuePair<string, Stream> testCase)
     {
         var demo = new DemoParser();
-        await demo.Start(testCase.Value, default);
+        await demo.ReadAllAsync(testCase.Value, default);
     }
 
     [Test]
-    public async Task Parse_AlternateBaseline()
+    public async Task ReadAll_AlternateBaseline()
     {
         var demo = new DemoParser();
-        await demo.Start(MatchmakingProtocol13968, default);
+        await demo.ReadAllAsync(MatchmakingProtocol13968, default);
     }
 }