+++ src/DemoFile/Model/DemoTick.cs
@@ -3,9 +3,9 @@ using System.Diagnostics;
 namespace DemoFile;
 
 /// <summary>
-/// Demo ticks start at 0 from the moment the demo starts recording.
-/// Some demo messages are recorded at DemoTick -1, a fake tick indicating
-/// that the messages occurred before the demo started recording.
+/// Demos start at <see cref="Zero"/> from the moment the demo starts recording.
+/// Note some demo messages are recorded at DemoTick -1 (<see cref="PreRecord"/>),
+/// a fake tick indicating that the messages occurred before the demo started recording.
 /// These include messages like <see cref="CSVCMsg_ServerInfo"/>.
 /// </summary>
 /// <param name="Value">Current demo tick. <c>-1</c> or greater.</param>
@@ -24,4 +24,9 @@ public readonly record struct DemoTick(int Value) : IComparable<DemoTick>
 
     public static DemoTick operator +(DemoTick tick, TimeSpan duration) => new((int)(tick.Value + duration.TotalSeconds * 64.0));
     public static DemoTick operator -(DemoTick tick, TimeSpan duration) => new((int)(tick.Value - duration.TotalSeconds * 64.0));
+
+    public static bool operator <(DemoTick left, DemoTick right) => left.Value < right.Value;
+    public static bool operator <=(DemoTick left, DemoTick right) => left.Value <= right.Value;
+    public static bool operator >(DemoTick left, DemoTick right) => left.Value > right.Value;
+    public static bool operator >=(DemoTick left, DemoTick right) => left.Value >= right.Value;
 }